[dev]サインインとサインアウト時のエラーハンドリングを記述#3
